+++ frontend/src/pages/Feed.jsx
@@ -1,21 +1,21 @@
-import { useEffect, useState, useCallback, useRef } from "react";
+import { useEffect, useState } from "react";
 import { Spinner } from "flowbite-react";
-import Post from "../components/Post";
+
 import useAuth from "../hooks/useAuth";
+import useScrollingFeed from "../hooks/useScrollingFeed";
+
+import Post from "../components/Post";
 import Composer from "../components/Composer";
 import SuggestedUsers from "../components/SuggestedUsers";
+
 import { getLatestFeed } from "../api/feeds.js";
 
 function Feed() {
   const auth = useAuth();
-  const [posts, setPosts] = useState([]);
   const [viewer, setViewer] = useState(null);
-  const [loading, setLoading] = useState(false);
-  const [hasMore, setHasMore] = useState(true);
-  const [page, setPage] = useState(0);
-  const POSTS_PER_PAGE = 5;
 
-  const firstLoad = useRef(true);
+  const { posts, feedLoading, hasMore, onPostDeleted, onPostCreated } =
+    useScrollingFeed({ POSTS_PER_PAGE: 5, fetchFeedFunction: getLatestFeed });
 
   useEffect(() => {
     if (auth.loggedIn) {
@@ -24,69 +24,6 @@ function Feed() {
       });
     }
   }, [auth]);
-
-  const fetchPosts = useCallback(
-    async (pageNumber) => {
-      if (loading) return;
-      setLoading(true);
-      try {
-        const response = await getLatestFeed({
-          limit: POSTS_PER_PAGE,
-          offset: POSTS_PER_PAGE * pageNumber,
-        });
-
-        if (response.success !== false) {
-          setPage(pageNumber + 1);
-          setPosts((prevPosts) => [...prevPosts, ...response.data]);
-          setHasMore(response.data.length === POSTS_PER_PAGE);
-          // Assuming there are more posts if the page is full. Could improve this
-        }
-      } catch (error) {
-        console.error("Error fetching posts:", error);
-      } finally {
-        setLoading(false);
-      }
-    },
-    [loading],
-  );
-
-  useEffect(() => {
-    if (firstLoad.current) {
-      fetchPosts(page);
-      firstLoad.current = false;
-    }
-  }, [fetchPosts, page]);
-
-  const handleScroll = useCallback(() => {
-    const mainContent = document.getElementById("main-content-scrollable");
-    if (
-      mainContent.scrollTop + mainContent.clientHeight >=
-      mainContent.scrollHeight - 100
-    ) {
-      console.log("Loading more posts...");
-      // Load more posts
-      if (!loading && hasMore) {
-        setLoading(true);
-        fetchPosts(page).finally(() => {
-          setLoading(false);
-        });
-      }
-    }
-  }, [fetchPosts, page, hasMore, loading]);
-
-  useEffect(() => {
-    const mainContent = document.getElementById("main-content-scrollable");
-    mainContent.addEventListener("scroll", handleScroll);
-    return () => mainContent.removeEventListener("scroll", handleScroll);
-  }, [handleScroll]);
-
-  const onPostDeleted = (postId) => {
-    setPosts((prevPosts) => prevPosts.filter((post) => post._id !== postId));
-  };
-
-  const onPostCreated = (post) => {
-    setPosts((prevPosts) => [post, ...prevPosts]);
-  };
 
   return (
     <div className="container mx-auto">
@@ -109,13 +46,13 @@ function Feed() {
             />
           ))}
 
-          {loading && (
+          {feedLoading && (
             <div className="p-4 text-center">
               <Spinner aria-label="Loading more posts" size="xl" />
             </div>
           )}
 
-          {!hasMore && !loading && (
+          {!hasMore && !feedLoading && (
             <div className="p-4 text-center text-gray-500">
               No more posts to load.
             </div>
